back/requirements.txt
django
psycopg2
django-cors-headers
djangorestframework
pykeepass
